Don't send confidence 0 until 3 times failed:
For some reason data from companion gets corrupted easily.
As we are displaying the bounding box in ground based on
confidence value, if the parsing went wrong we set confidence
to 0. This lead to flickering rectangle, as every now and
then the message sent is not complete or corrupt. Now we only
force confidence = 0 if 3 times the message has been corrupted,
( more than 36 bits unexpected )