Wall mode: update after rebase to master:
The whole AC_Avoid was modified as well
to work in 3D, so some changes were re-
quired here.

This is the wall mode feature, activating
it makes the copter move towards the
closest obstacle until the margin, so it
can somehow follow the wall.

-Arducopter/GCS_mavlink and RC_Channel:
links for activating this by mavlink

and by rc switch.

-AC_Avoid: all the changes for actually
moving towards obstacle is made here