auto flight mode switch and ekf source:
We have a couple new global variables in copter.h
and a new scheduler fucntion in usercode.cpp.

if we click in loiter, it will switch automatically
between alt hold and loiter based on tracking quality.
if medium or lower will go to alt hold, if high to loiter.

This will follow the transition times given by the parameters
aut_time, so brief changes don't actually change mode.

Also, in this commit we change to src0 for loiter, and src1 for
all the rest of flight modes